split built in resolver tests to individual files
